Split offspringscoring.py into two files (other = variantselection.py).
