+++ whatshap/cli/polyphasegenetic.py
@@ -19,13 +19,8 @@ from whatshap import __version__
 from whatshap.core import ClusterEditingSolver, Read, ReadSet
 from whatshap.cli import log_memory_usage, CommandLineError
 from whatshap.polyphaseplots import create_genetic_plots
-from whatshap.offspringscoring import (
-    get_variant_scoring,
-    compute_phasable_variants,
-    filter_variants,
-    get_offspring_gl,
-    correct_variant_types,
-)
+from whatshap.variantselection import compute_phasable_variants, filter_variants
+from whatshap.offspringscoring import get_offspring_gl, correct_variant_types, get_variant_scoring
 from whatshap.timer import StageTimer
 from whatshap.vcf import VcfReader, PhasedVcfWriter, PloidyError
 from whatshap.clusterarrangement import arrange_clusters
